+++ release-notes.md
@@ -2,3 +2,7 @@
 =====
 	[add] define basic application settings 
 	[add] parse applications settings using MatthiWare.CommandLineParser
+1.0.2
+=====
+	[add] usage of the library SixLabors.ImageSharp for image handling
+	[add] application resize logic